[ci skip] use static_cast avoid warnings

+++ applications/mne_anonymize/fiffanonymizer.cpp
@@ -179,15 +179,15 @@ FiffAnonymizer::FiffAnonymizer(const FiffAnonymizer& obj)
 , m_printInSameLineHelper(qDebug())
 , m_bPrintInSameLine(true)
 {
-    
+
     //    m_BDfltMAC.resize(8);
     //    memcpy(m_BDfltMAC.data(),obj.m_BDfltMAC.data(),8);
-    
+
     m_pBlockTypeList = QSharedPointer<QStack<int32_t> >(new QStack<int32_t>(*obj.m_pBlockTypeList));
     //    m_pBlockTypeList->resize(obj.m_pBlockTypeList->size());
     //    memcpy(m_pBlockTypeList->data(),obj.m_pBlockTypeList.data(),
     //           static_cast<size_t>(obj.m_pBlockTypeList->size()));
-    
+
     m_pOutDir = QSharedPointer<QVector<FiffDirEntry> >(new QVector<FiffDirEntry>);
     //    m_pOutDir->resize(obj.m_pOutDir->size());
     //    memcpy(m_pOutDir->data(),obj.m_pOutDir->data(),
@@ -240,7 +240,7 @@ FiffAnonymizer::FiffAnonymizer(FiffAnonymizer &&obj)
 {
     //    m_BDfltMAC.resize(8);
     //    memcpy(m_BDfltMAC.data(),obj.m_BDfltMAC.data(),8);
-    
+
     m_pBlockTypeList.swap(obj.m_pBlockTypeList);
     m_pOutDir.swap(obj.m_pOutDir);
 }
@@ -259,7 +259,7 @@ int FiffAnonymizer::anonymizeFile()
     printIfVerbose("Version: " + versionStr);
     printIfVerbose("Current date: " + QDateTime::currentDateTime().toString("ddd MMMM d yyyy hh:mm:ss"));
     printIfVerbose(" ");
-    
+
     FiffStream inStream(&m_fFileIn);
     printIfVerbose("Input file: " + m_fFileIn.fileName());
     if(inStream.open(QIODevice::ReadOnly))
@@ -269,7 +269,7 @@ int FiffAnonymizer::anonymizeFile()
         qCritical() << "FiffAnonymizer::run - Problem opening the input file: " << m_fFileIn.fileName();
         return 1;
     }
-    
+
     FiffStream outStream(&m_fFileOut);
     printIfVerbose("Output file: " + m_fFileOut.fileName());
     if(outStream.device()->open(QIODevice::ReadWrite))
@@ -279,19 +279,19 @@ int FiffAnonymizer::anonymizeFile()
         qCritical() << "FiffAnonymizer::run - Problem opening the output file: " << m_fFileOut.fileName();
         return 1;
     }
-    
+
     FiffTag::SPtr pInTag = FiffTag::SPtr::create();
     FiffTag::SPtr pOutTag = FiffTag::SPtr::create();
-    
+
     inStream.read_tag(pInTag,0);
     //    FiffTag::convert_tag_data(pInTag,FIFFV_BIG_ENDIAN,FIFFV_NATIVE_ENDIAN);
-    
+
     //info in a tag FIFF_COMMENT (206) depends on the type of block it is in. Therefore, in order to
     //anonymize it we not only have to know the kind of the current tag, but also which type of block
     //we're currently in. BlockTypeList is a stack container used to keep track of this.
     m_pBlockTypeList->clear();
     updateBlockTypeList(pInTag);
-    
+
     printIfVerbose("Reading info in the file.");
     if(checkValidFiffFormatVersion(pInTag))
     {
@@ -303,11 +303,11 @@ int FiffAnonymizer::anonymizeFile()
     }
     censorTag(pOutTag,pInTag);
     pOutTag->next = FIFFV_NEXT_SEQ;
-    
+
     //we build the tag directory on the go
     addEntryToDir(pOutTag,outStream.device()->pos());
     outStream.write_tag(pOutTag);
-    
+
     while(pInTag->next != -1)
     {
         inStream.read_tag(pInTag);
@@ -322,37 +322,37 @@ int FiffAnonymizer::anonymizeFile()
         addEntryToDir(pOutTag,outStream.device()->pos());
         outStream.write_tag(pOutTag);
     }
-    
+
     if(inStream.close())
     {
         printIfVerbose("Input file finished. All tags have been correctly anonymized.");
     } else {
         qCritical() << "FiffAnonymizer::run - Problem closeing the input file: " << m_fFileIn.fileName();
     }
-    
+
     addFinalEntryToDir();
     fiff_long_t posOfDirectory(outStream.device()->pos());
     writeDirectory(&outStream);
     updatePointer(&outStream,FIFF_DIR_POINTER,posOfDirectory);
     updatePointer(&outStream,FIFF_FREE_LIST,-1);
-    
+
     if(outStream.close())
     {
         printIfVerbose("Input file finished. All tags have been correctly anonymized.");
     } else {
         qCritical() << "FiffAnonymizer::run - Problem closing the output file: " << m_fFileOut.fileName();
     }
-    
+
     if(checkDeleteInputFile())
     {
         deleteInputFile();
     }
-    
+
     if(checkRenameOutputFile())
     {
         renameOutputFileAsInputFile();
     }
-    
+
     printIfVerbose(" ");
     printIfVerbose("----------------------------------------------------------------------------");
     printIfVerbose(" ");
@@ -384,8 +384,8 @@ bool FiffAnonymizer::checkValidFiffFormatVersion(FiffTag::SPtr pTag)
     if(pTag->kind == FIFF_FILE_ID)
     {
         FiffId fileId = pTag->toFiffID();
-        int inMayorVersion = (fileId.version & 0xFFFF0000) >> 16;
-        int inMinorVersion = (fileId.version & 0x0000FFFF);
+        int inMayorVersion = (static_cast<uint32_t>(fileId.version) & 0xFFFF0000) >> 16;
+        int inMinorVersion = (static_cast<uint32_t>(fileId.version) & 0x0000FFFF);
         double inVersion = inMayorVersion + inMinorVersion/10.;
         if(inVersion > maxValidFiffVerion)
             return false;
@@ -430,7 +430,7 @@ void FiffAnonymizer::dir2tag(FiffTag::SPtr pTag)
     pTag->next = -1;
     pTag->resize(m_pOutDir->size() * 4  * 4);
     pTag->clear();
-    
+
     if(m_pOutDir->size() > 0)
     {
         QByteArray pInt8(sizeof(fiff_int_t),0);
@@ -461,7 +461,7 @@ void FiffAnonymizer::writeDirectory(FiffStream* stream,
         if(file)
             stream->device()->seek(file->size());
     }
-    
+
     *stream << static_cast<quint32>(FIFF_DIR);
     *stream << static_cast<quint32>(FIFFT_DIR_ENTRY_STRUCT);
     *stream << static_cast<quint32>
@@ -484,7 +484,7 @@ void FiffAnonymizer::updatePointer(FiffStream* stream,
                                    fiff_long_t newPos)
 {
     fiff_long_t tagInfoSize = 16;
-    
+
     for(int i=0; i < m_pOutDir->size(); ++i) {
         if(m_pOutDir->at(i).kind != tagKind) {
             stream->device()->seek(m_pOutDir->at(i).pos+tagInfoSize);
@@ -499,7 +499,7 @@ void FiffAnonymizer::updatePointer(FiffStream* stream,
 
 bool FiffAnonymizer::checkDeleteInputFile()
 {
-    
+
     if(m_bDeleteInputFileAfter) //false by default
     {
         qDebug() << "You have requested to delete the input file: " + m_fFileIn.fileName();
@@ -518,7 +518,7 @@ bool FiffAnonymizer::checkDeleteInputFile()
         {
             return true;
         }
-        
+
     }
     return false;
 }
@@ -529,11 +529,11 @@ bool FiffAnonymizer::checkDeleteInputFile()
 int FiffAnonymizer::censorTag(FiffTag::SPtr outTag,FiffTag::SPtr inTag)
 {
     int sizeDiff(0);
-    
+
     outTag->kind = inTag->kind;
     outTag->type = inTag->type;
     outTag->next = inTag->next;
-    
+
     switch (inTag->kind)
     {
     //all these 'kinds' of tags contain a fileID struct, which contains info related to
@@ -580,14 +580,14 @@ int FiffAnonymizer::censorTag(FiffTag::SPtr outTag,FiffTag::SPtr inTag)
     {
         QDateTime inMeasDate(QDateTime::fromSecsSinceEpoch(*inTag->toInt()).date());
         QDateTime outMeasDate;
-        
+
         if(m_bUseMeasurementDayOffset)
         {
             outMeasDate = inMeasDate.addDays(-m_iMeasurementDayOffset);
         } else {
             outMeasDate = m_dateMeasurmentDate;
         }
-        
+
         fiff_int_t outData[1];
         outData[0]=static_cast<int32_t>(outMeasDate.toSecsSinceEpoch());
         memcpy(outTag->data(),reinterpret_cast<char*>(outData),1);
@@ -660,14 +660,14 @@ int FiffAnonymizer::censorTag(FiffTag::SPtr outTag,FiffTag::SPtr inTag)
     {
         QDateTime inBirthday(QDate::fromJulianDay(*inTag->toJulian()));
         QDateTime outBirthday;
-        
+
         if(m_bUseSubjectBirthdayOffset)
         {
             outBirthday = inBirthday.addDays(-m_iMeasurementDayOffset);
         } else {
             outBirthday = m_dateSubjectBirthday;
         }
-        
+
         fiff_int_t outData[1];
         outData[0] = static_cast<int32_t> (outBirthday.toSecsSinceEpoch());
         memcpy(outTag->data(),reinterpret_cast<char*>(outData),1);
@@ -772,11 +772,11 @@ int FiffAnonymizer::censorTag(FiffTag::SPtr outTag,FiffTag::SPtr inTag)
         outTag->resize(inTag->size());
         memcpy(outTag->data(),inTag->data(),static_cast<size_t>(inTag->size()));
     }
-        
+
     }
-    
+
     sizeDiff = outTag->size() - inTag->size();
-    
+
     return sizeDiff;
 }
 
@@ -953,7 +953,7 @@ bool FiffAnonymizer::checkRenameOutputFile()
                 deleteInputFile();
                 return true;
             } else {
-                
+
                 qCritical() << " ";
                 qCritical() << "You have requested to save the output file with the same name as the input file.";
                 qCritical() << "This cannot be done without deleting or modifying the input file.";
